Include RootHistV7ToV6Conversion better
* merged hpp.dcl and hpp
* rename files
* change extenisons
* added partial Test
